get rid of duplicate tests with shared module
