+++ spec/model/range_methods_spec.rb
@@ -1,66 +1,15 @@
 require "spec_helper"
 
 RSpec.describe PgParty::Model::RangeMethods do
-  let(:key) { "created_at::date" }
   let(:decorator) { instance_double(PgParty::ModelDecorator) }
 
   before do
-    model.range_partition_by(key)
     allow(PgParty::ModelDecorator).to receive(:new).with(model).and_return(decorator)
   end
 
   subject(:model) do
     Class.new do
-      extend PgParty::Model::Methods
-    end
-  end
-
-  describe ".partition_key" do
-    subject { model.partition_key }
-
-    it { is_expected.to eq("created_at::date") }
-  end
-
-  describe ".partition_column" do
-    subject { model.partition_column }
-
-    context "when key has cast" do
-      it { is_expected.to eq("created_at") }
-    end
-
-    context "when key does not have cast" do
-      let(:key) { "created_at" }
-
-      it { is_expected.to eq("created_at") }
-    end
-  end
-
-  describe ".partition_cast" do
-    subject { model.partition_cast }
-
-    context "when key has cast" do
-      it { is_expected.to eq("date") }
-    end
-
-    context "when key does not have cast" do
-      let(:key) { "created_at" }
-
-      it { is_expected.to be_nil }
-    end
-  end
-
-  describe ".cached_partitions" do
-    subject { model.cached_partitions }
-
-    it { is_expected.to be_nil }
-  end
-
-  describe ".partitions" do
-    subject { model.partitions }
-
-    it "delegates to decorator" do
-      expect(decorator).to receive(:partitions)
-      subject
+      extend PgParty::Model::RangeMethods
     end
   end
 
@@ -89,17 +38,6 @@ RSpec.describe PgParty::Model::RangeMethods do
 
     it "delegates to decorator" do
       expect(decorator).to receive(:range_partition_key_in).with(start_range, end_range)
-      subject
-    end
-  end
-
-  describe ".partition_key_eq" do
-    let(:value) { Date.current }
-
-    subject { model.partition_key_eq(value) }
-
-    it "delegates to decorator" do
-      expect(decorator).to receive(:partition_key_eq).with(value)
       subject
     end
   end
