+++ spec/model/shared_methods_spec.rb
@@ -1,0 +1,35 @@
+require "spec_helper"
+
+RSpec.describe PgParty::Model::SharedMethods do
+  let(:decorator) { instance_double(PgParty::ModelDecorator) }
+
+  before do
+    allow(PgParty::ModelDecorator).to receive(:new).with(model).and_return(decorator)
+  end
+
+  subject(:model) do
+    Class.new do
+      extend PgParty::Model::SharedMethods
+    end
+  end
+
+  describe ".partitions" do
+    subject { model.partitions }
+
+    it "delegates to decorator" do
+      expect(decorator).to receive(:partitions)
+      subject
+    end
+  end
+
+  describe ".partition_key_eq" do
+    let(:value) { Date.current }
+
+    subject { model.partition_key_eq(value) }
+
+    it "delegates to decorator" do
+      expect(decorator).to receive(:partition_key_eq).with(value)
+      subject
+    end
+  end
+end
